+++ src/Exception/InvalidConnectionPurgeModeException.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Kununu\DataFixtures\Purger;
+namespace Kununu\DataFixtures\Exception;
 
 final class InvalidConnectionPurgeModeException extends \RuntimeException
 {
