+++ tests/Purger/CachePoolPurgerTest.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace Kununu\DataFixtures\Tests\Purger;
 
+use Kununu\DataFixtures\Exception\PurgeFailedException;
 use Kununu\DataFixtures\Purger\CachePoolPurger;
-use Kununu\DataFixtures\Purger\PurgeFailedException;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Psr\Cache\CacheItemPoolInterface;
